delete from all relays

+++ src/main.rs
@@ -6,6 +6,9 @@ use toml;
 use regex::Regex;
 use clap::Parser;
 use nostr_sdk::prelude::*;
+use std::collections::HashSet;
+use std::time::{Duration, Instant};
+use tokio::select;
 
 #[derive(Parser, Debug)]
 #[command(version, about, long_about = None)]
@@ -67,12 +70,12 @@ fn validate_content(content: &String) -> Result<()> {
 
     Ok(())
 }
-async fn publish_article(file_name: String, article_identifier: String, title: Option<String>, image: Option<Url>, summary: Option<String>, published_at: Option<u64>, client: Client) -> Result<()> {
+async fn publish_article(file_name: String, article_identifier: String, title: Option<String>, image: Option<Url>, summary: Option<String>, published_at: Option<u64>, client: Client, public_key: PublicKey) -> Result<()> {
     let content = fs::read_to_string(file_name).with_context(|| format!("Content file could not be read."))?;
     validate_content(&content)?;
     
     let mut tags = Vec::from([
-        Tag::identifier(article_identifier),
+        Tag::identifier(article_identifier.clone()),
     ]);
 
     if let Some(title) = title {
@@ -90,25 +93,85 @@ async fn publish_article(file_name: String, article_identifier: String, title: O
 
     let timestamp = published_at.unwrap_or_else(|| Timestamp::now().as_u64());
     tags.push(Tag::from_standardized(TagStandard::PublishedAt(Timestamp::from(timestamp))));
-     
+
+    let coordinate = Coordinate { kind: Kind::LongFormTextNote, public_key: public_key, identifier: article_identifier };
+    tags.push(Tag::from_standardized(TagStandard::Coordinate { coordinate: coordinate, relay_url: None, uppercase: false }));
+    
     // Publish a text note
     let builder = EventBuilder::long_form_text_note(content).tags(tags);
     let event = client.send_event_builder(builder).await?;
     println!("Generated EventId: {}", event.to_hex());
-
+    // let event: Event = client.sign_event_builder(builder).await?;
+    // println!("{:?}", event.tags);
+    // println!("{:?}", event.coordinate());
     Ok(())
 }
 
 async fn delete_article(article_identifier: String, client: Client, public_key: PublicKey) -> Result<()> {
     let coordinate = Coordinate { kind: Kind::LongFormTextNote, public_key: public_key, identifier: article_identifier };
+
+    // Create subscription to find the article to delete
+    let subscription = Filter::new()
+        .author(public_key)
+        .kind(Kind::LongFormTextNote)
+        .coordinate(&coordinate);
+    let Output { val: sub_id_1, .. } = client.subscribe(subscription, None).await?;
+
+    let mut event_ids: HashSet<EventId> = HashSet::new();
+    let mut eose_received = false;
+
+    let timeout_duration = Duration::from_secs(10);
+    let start = Instant::now();
+
+    let mut notifications = client.notifications();
+
+    loop {
+        if eose_received || start.elapsed() > timeout_duration {
+            break;
+        }
+
+        select! {
+            Ok(notification) = notifications.recv() => {
+                match notification {
+                    RelayPoolNotification::Event { event, subscription_id: sid, .. } if sid == sub_id_1 => {
+                        event_ids.insert(event.id);
+                    }
+                    RelayPoolNotification::Message { message, .. } => {
+                        if let RelayMessage::EndOfStoredEvents(sid) = message {
+                            if sid.into_owned() == sub_id_1 {
+                                eose_received = true;
+                            }
+                        }
+                    }
+                    _ => {},
+                }
+            }
+            _ = tokio::time::sleep(timeout_duration - start.elapsed()) => {
+                break;
+            }
+        }
+    }
+
+    client.unsubscribe(&sub_id_1).await;
+
+    if event_ids.is_empty() {
+        println!("Not found events to delete.");
+        return Ok(());
+    }
+
+    // Create deletion event
     let request = EventDeletionRequest::new()
                                             .coordinate(coordinate)
                                             .reason("Deleted by user request");
     let k_tag = Tag::from_standardized(TagStandard::Kind { kind: Kind::LongFormTextNote  , uppercase: false });
-    let builder = EventBuilder::delete(request)
+    let mut builder = EventBuilder::delete(request)
                                     .tag(k_tag);
+
+    for id in &event_ids {
+        builder = builder.tag(Tag::event(*id));
+    }
     let event = client.send_event_builder(builder).await?;
-    println!("Deletetion EventId: {}", event.to_hex());
+    println!("Deletion EventId: {}", event.to_hex());
     // let event: Event = client.sign_event_builder(builder).await?;
     // println!("{:?}",event.tags);
     Ok(())
@@ -144,7 +207,7 @@ async fn main() -> Result<()> {
     client.connect().await;
 
     match args.command {
-        Commands::Publish { file_name, article_identifier, title, image, summary, published_at } => { publish_article(file_name, article_identifier, title, image, summary, published_at, client).await? },
+        Commands::Publish { file_name, article_identifier, title, image, summary, published_at } => { publish_article(file_name, article_identifier, title, image, summary, published_at, client, keys.public_key()).await? },
         Commands::Delete { article_identifier } => { delete_article(article_identifier, client, keys.public_key()).await? }
     }
